requirements.txt
absl-py==0.8.1
alabaster==0.7.12
allennlp==0.9.0
asn1crypto==0.24.0
astor==0.8.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
bcolz==1.2.1
beautifulsoup4==4.7.1
bleach==3.1.0
blessings==1.7
blis==0.2.4
bokeh==2.1.1
boto3==1.10.34
botocore==1.13.34
Bottleneck==1.2.1
certifi==2020.6.20
cffi==1.12.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.2.1
conllu==1.3.1
cryptography==2.7
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
dask==2.3.0
dataclasses==0.6
decorator==4.4.0
defusedxml==0.6.0
dill==0.2.9
distributed==2.5.2
docutils==0.15.2
dominate==2.4.0
editdistance==0.5.3
entrypoints==0.3
fastprogress==0.1.21
flaky==3.6.1
Flask==1.0.3
Flask-Cors==3.0.8
fsspec==0.8.0
ftfy==5.6
gast==0.2.2
gevent==1.4.0
google-pasta==0.1.7
gpustat==0.6.0
greenlet==0.4.15
grpcio==1.24.1
h5py==2.10.0
HeapDict==1.0.1
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==1.2.0
ipykernel==5.1.1
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
json5==0.8.4
jsonnet==0.14.0
jsonpickle==1.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.0.2
jupyterlab-server==1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
locket==0.2.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
mkl-service==2.0.2
more-itertools==8.0.2
msgpack==0.6.1
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.5
notebook==6.0.0
numexpr==2.6.9
numpy==1.16.4
numpydoc==0.9.1
nvidia-ml-py3==7.352.0
olefile==0.46
opencv-contrib-python==4.1.0.25
opt-einsum==3.1.0
overrides==2.5
packaging==19.0
pandas==0.25.0
pandocfilters==1.4.2
parsimonious==0.8.1
parso==0.5.0
partd==1.1.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
plac==0.9.6
pluggy==0.13.1
preshed==2.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.10.0
psutil==5.7.2
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.14.11
PySocks==1.7.0
pytest==5.3.1
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1.1
pyzmq==18.0.0
qtconsole==4.5.2
regex==2018.1.10
requests==2.22.0
responses==0.10.7
s3transfer==0.2.1
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
Send2Trash==1.5.0
sentencepiece==0.1.83
six==1.12.0
snowballstemmer==2.0.0
sortedcontainers==2.2.2
soupsieve==1.9.2
spacy==2.1.8
Sphinx==2.2.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
srsly==0.1.0
tables==3.4.4
tblib==1.7.0
tensorboard==2.0.0
tensorboardX==1.9
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
thinc==7.0.8
toolz==0.10.0
torch==1.2.0
torchfile==0.1.0
torchvision==0.2.1
tornado==6.0.3
tqdm==4.32.1
traitlets==4.3.2
typing==3.6.4
typing-extensions==3.7.4.3
ujson==1.35
Unidecode==1.1.1
urllib3==1.24.2
visdom==0.1.8.8
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.5
widgetsnbextension==3.5.0
word2number==1.1
wrapt==1.11.2
yacs==0.1.6
zict==2.0.0
zipp==0.6.0
